Patch for Clang
Clang correctly diagnoses that we cannot use 'constexpr' to define
a function-local static 'auto' object of a lambda expression type
as the constructors are not literal constructors.  g++ supports this
as an extension, but it is not portable, and not clear that wg21
would be interested in adding such an extension.