Document issues statuses better
We have added a number of issue statuses to the issues list
processing software since the text of the standing documents
was last reviewed, and quite a few statuses are not yet
described.  This re-orders the status descriptions into
something approaching a workflow, with the closed statuses
described in order of published document (most recent first)
followed by various NAD and other administratively closed
cases.