Revised ordering of revision list
Th summary of status-changes in the revision list for the three standing
documents is now ordered by status, using the same order as the order-by-
status view, listing the changed-to status first, to better identift the
grouping.  The 'from' status in now parenthetical, as it seems to make it
easier to pick out the two different status strings this way.

In addition, the documents are now created in a 'mailing' subdirectory,
as this simplifies the publishing process for AJM.  This change is
provisional, waiting on feedback from the backup list maintainers.