Factor functionality into separate files
Start refactoring the main lists.cpp file into more focussed components.

The main program is growing in functionality and it becomes difficult
to find and track the code responsible for any given functionality.
Separating out more focussed components will give an opportunity to
better document contracts in headers, and gain a higher-level view
of the code.

This initial refactoring is a little coarse, as it is based on the
existing code dependencies, rather than a real factoring of concerns,
but it does give a good first cut that can be refined.