factor out document formatting functions
Create a new file for 'report generation' that manages all the
functions reposnible for preparing and writing the various
documents supported by the issues list software.  The next
step is to make these methods of a report_generator object
that can store persistent info like the various databases
that we have loaded into memory, rather than passing them
as parameters to each call.

We also want to switch to writing to a stream, rather than
depending on <fstream> as an implementation detail inside
the class.

Finally, note that this change-set conflates a switch
from 'std::pair' to 'std::tuple' for nothing other
than didactic purposes.  I have a preference to act
as if 'pair' never existed and use the more general
'tuple' in code susceptible to frequnet change.