Extract section number wrapper-type
The section-number is another independant piece useed to assemble the
LWG documents.  In this case, a section number is the numbered clause
in the relevant document, and a named reference to that document,
where an empty string implies the main standard itself.

This extraction leaves the main 'lists.cpp' file small enough that
we can finally remove the 'using namespace lwg' and be explicit
about the types and functions specific to the issues list software.