Preparing additional reports
Work towards a tool that will create a single document to
pass to the project editor after a meeting containing only
the resolutions of all issues adopted at that meeting, in
section order.

Some additional maintenance on the reports that generate
papers on Immediate issues to resolve during a meeting,
based on feedback from Clark Nelson on better ISO form
of the resulting papers.