Simplified setting status through bash scripts
The common pattern to move status on a block of issues is now:
   list_issues OLD_STATUS | set_status NEW_STATUS
This hits a problem when NEW_STATUS has embedded whitespace,
so the work-around for now is to turn whitespace into under-
scores, and reverse the transformation inside the set_status
tool.  This should be possible through appropriate use of
quotes and similar in bash, but has proved the better of me
so far, so opting for the quick fix.  Suspect there may be
similar problems with handling the Windows command line as
well.