Shell script to aid mass-status-updates
This script allows us to pass the issue number argument resulting
from a list_issues query as the first parameter to set_status. In
hindsight it might have been easier to re-order arguments in the
set_status program, but this is a quick-and-easy fix before we
make the command-line parsing of our tools smarter.